refactor: add tabs layout and move screens
