+++ react-native-blueye/app/_layout.jsx
@@ -92,80 +92,6 @@ export default function Layout() {
                 />
               </View>
 
-              {/* Bottom Tabs */}
-              <View className="fixed bottom-0 left-0 right-0 bg-phase2Cards dark:bg-phase2CardsDark border-t border-phase2Borders dark:border-phase2BordersDark">
-                <View className="flex-row">
-                  {/* Botón Inicio/Mapa */}
-                  <Link
-                    href="/"
-                    className={`flex flex-1 flex-col py-2 items-center text-center ${
-                      currentRoute === "/"
-                        ? "bg-phase2Cards dark:bg-phase2CardsDark text-phase2Titles dark:text-phase2TitlesDark"
-                        : "bg-phase2Buttons dark:bg-phase2ButtonsDark text-white hover:bg-phase2Borders dark:hover:bg-phase2BordersDark"
-                    }`}
-                  >
-                    <MaterialCommunityIcons
-                      name="map"
-                      size={28}
-                      color={
-                        currentRoute === "/"
-                          ? colorScheme === "dark"
-                            ? "rgb(230, 230, 250)" // phase2TitlesDark
-                            : "rgb(30, 30, 60)" // phase2Titles
-                          : "white"
-                      }
-                    />
-
-                    <Text className="text-sm">Mapa</Text>
-                  </Link>
-
-                  {/* Botón Chat-AI */}
-                  <Link
-                    href="/chat-ai"
-                    className={`flex flex-1 flex-col py-2 items-center text-center ${
-                      currentRoute === "/chat-ai"
-                        ? "bg-phase2Cards dark:bg-phase2CardsDark text-phase2Titles dark:text-phase2TitlesDark"
-                        : "bg-phase2Buttons dark:bg-phase2ButtonsDark text-white hover:bg-phase2Borders dark:hover:bg-phase2BordersDark"
-                    }`}
-                  >
-                    <MaterialCommunityIcons
-                      name="robot-happy"
-                      size={28}
-                      color={
-                        currentRoute.startsWith("/chat-ai")
-                          ? colorScheme === "dark"
-                            ? "rgb(230, 230, 250)" // phase2TitlesDark
-                            : "rgb(30, 30, 60)" // phase2Titles
-                          : "white"
-                      }
-                    />
-                    <Text className="text-sm">Chat-AI</Text>
-                  </Link>
-
-                  {/* Botón Configuración */}
-                  <Link
-                    href="/alerts"
-                    className={`flex flex-1 flex-col py-2 items-center text-center justify-center ${
-                      currentRoute === "/alerts"
-                        ? "bg-phase2Cards dark:bg-phase2CardsDark text-phase2Titles dark:text-phase2TitlesDark"
-                        : "bg-phase2Buttons dark:bg-phase2ButtonsDark text-white hover:bg-phase2Borders dark:hover:bg-phase2BordersDark"
-                    }`}
-                  >
-                    <MaterialCommunityIcons
-                      name="bell-alert-outline"
-                      size={28}
-                      color={
-                        currentRoute.startsWith("/alerts")
-                          ? colorScheme === "dark"
-                            ? "rgb(230, 230, 250)" // phase2TitlesDark
-                            : "rgb(30, 30, 60)" // phase2Titles
-                          : "white"
-                      }
-                    />
-                    <Text className="text-sm">Alertas</Text>
-                  </Link>
-                </View>
-              </View>
             </View>
           </TamaguiProvider>
         </SafeAreaProvider>
